DOCS: Configuration section - cleanup
As per discussion in the OSA docs summit session, clean up
of installation guide. This fixes typos, minor RST mark up
changes, and passive voice.

This patch also merges a some of the sections into the larger
chapter. This is in an effort to remove multiple smaller
files.

This patch is the first of many to avoid major conflicts.

Combined backport of:
 - I968f9350db8c4e2450fc9d7d59de54bf75d09c53
   (cherry picked from commit 3107fdd298fd90b02153ebca8703f4357ce99c5f)
 - I2b1582812e638e2b3b455b7c34b93d13e08a168a
   (cherry picked from commit 55155f301e13e5c0867ba4a42fc714dfdc6f72f2)

Change-Id: I2b1582812e638e2b3b455b7c34b93d13e08a168a