[GATE UNBLOCKER] Lock CPU map script to a release
The CPU map script from devstack has been removed upstream.
This needs to be locked down to ensure we're not using the head
of master when depending on this script.

Upstream commit where the script was removed: [0].

To stablize the gate the script has been locked down using the a
SHA from the "stable/mitaka" branch within upstream "devstack/tools" [1].

[0] - https://github.com/openstack-dev/devstack/commit/214459cdf776a77dacd70a9205c5b8a6ceaff31a
[1] - http://git.openstack.org/cgit/openstack-dev/devstack/plain/tools/cpu_map_update.py?h=0181c486f7e0126c4f46afd4dbc789b8fe096902

Change-Id: I0d55e799e4b4caff502438747fc65729dcaefd7e
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>